rearranging middleware so it's all in one file
