+++ server/db.js
@@ -1,0 +1,11 @@
+const Pool = require("pg").Pool;
+
+const pool = new Pool({
+  user: process.env.USER,
+  password: process.env.PASSWORD,
+  host: process.env.HOST,
+  port: process.env.DBPORT,
+  database: process.env.DATABASE,
+});
+
+module.exports = pool;
